backend/requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
greenlet==3.1.1
html5tagger==1.3.0
httptools==0.6.4
multidict==6.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sanic==24.12.0
sanic-routing==23.12.0
SQLAlchemy==2.0.36
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.21.0
websockets==14.1
